+++ src/index.ts
@@ -1,3 +1,4 @@
+export * from './app-check/index.js';
 export * from './firebase/index.js';
 export * from './identity-platform/index.js';
 export * from './logging/index.js';