+++ CHANGELOG.md
@@ -12,3 +12,4 @@ Features:
 - Define the logger configuration for GCP.
 - Implement the `PubSubPublisher`, the `PubSubEventHandlerInterceptor`, and the corresponding modules.
 - Implement the `PubSubHealthIndicator`.
+- Implement the `AppCheckGuard`.